+++ src/main/java/com/dhulai/repository/ProductRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dhulai.entity.Products;
 
-public interface ProductRepository extends JpaRepository<Products,Integer>{
+public interface ProductRepository extends JpaRepository<Products,Long>{
 
     
 }
